+++ public/javascripts/game_partials/_1_init.js
@@ -1,64 +1,3 @@
-var stage, w, h, loader;
-var sky, sun, clouds, road, buildings, backBg, frontBg, ambulance, speed = 100,
-  createTreeStrip, addTrees, treeStrip, ditch, buildings, tokens, sound, flag = true;
-var score = {
-  value: 0,
-  ob: {}
-};
-
-
-stage = new createjs.Stage('game-holder');
-
-// get canvas width and height for later calculations:
-w = stage.canvas.width;
-h = stage.canvas.height;
-
-var IMAGES_HOLDER = [{
-  src: "Sky.png",
-  id: "sky"
-}, {
-  src: "Sun.png",
-  id: "sun"
-}, {
-  src: "Buildings-Back.png",
-  id: "backBg"
-}, {
-  src: "Buildings-Front.png",
-  id: "frontBg"
-}, {
-  src: "Road.png",
-  id: "road"
-}, {
-  src: "Clouds.png",
-  id: "clouds"
-}, {
-  src: "tree-sprite.png",
-  id: "tree"
-}, {
-  src: "Building-1.png",
-  id: "building1"
-}, {
-  src: "Building-2.png",
-  id: "building2"
-}, {
-  src: "Building-3.png",
-  id: "building3"
-}, {
-  src: "ambulance-sprite.png",
-  id: "amb"
-}, {
-  src: "Ditch.png",
-  id: "ditch"
-}, {
-  src: "Token-1.png",
-  id: "token1"
-}, {
-  src: "Token-2.png",
-  id: "token2"
-}
-];
-
-loader = new createjs.LoadQueue(false);
 $(document).on('initialize-game', function() {
 
 loader.addEventListener("complete", handleComplete);
@@ -255,91 +194,3 @@ function handleComplete(e) {
 };
 
 });
-function initTweens(params) {
-    createjs.Tween.get(road, {
-        loop: true
-    }).to({
-        x: -w
-    }, speed * 45);
-
-    createjs.Tween.get(clouds, {
-        loop: true
-    }).to({
-        x: w
-    }, speed * 5000);
-
-    createjs.Tween.get(backBg, {
-        loop: true
-    }).to({
-        x: -w
-    }, speed * 500);
-
-    createjs.Tween.get(frontBg, {
-        loop: true
-    }).to({
-        x: -w
-    }, speed * 80);
-}
-function tickHandler(event) {
-    var deltaS = event.delta;
-    var maxSpeed = 1000;
-    var fSpeed = deltaS / 5;
-
-    treeStrip.forEach(function (tree, treeIndex) {
-        treeBounds = tree.getBounds();
-        tree.x = ((tree.x + treeBounds.width) <= 0) ? tree.x = w + treeBounds.width + (Math.random() * w) : tree.x - fSpeed;
-    });
-
-
-    buildings.forEach(function (building, itsIndex) {
-        buildingBounds = building.getBounds();
-        building.x = ((building.x + buildingBounds.width) <= 0) ? building.x = w + buildingBounds.width + (Math.random() * w) : building.x - fSpeed;
-    });
-
-    ditch.x = ((ditch.x + 200) < 0) ? w + (0.5 + Math.random()) * w : ditch.x - fSpeed;
-
-    // Check if collsion has occured
-    var pt = ditch.localToLocal(0, 0, ambulance);
-    if (ambulance.hitTest(pt.x, pt.y)) {
-        ambulance.gotoAndPlay("hickup");
-    }
-
-    // Randomly drop tokens
-    if (Math.random() * 1000 > 980) {
-        dropTokens();
-    }
-
-    // Check if token collected
-    tokens.forEach(function (token) {
-        var pt = token.localToLocal(20, 0, ambulance);
-        if (ambulance.hitTest(pt.x, pt.y)) {
-            tokenCollected(token);
-        }
-    }, this);
-
-    stage.update(event);
-}
-function moveSprite(e) {
-    // 39 - right arrow
-    // 37 - left arrow
-    if (e.keyCode === 39 || e.keyCode === 37) {
-
-        if (e.keyCode === 39) {
-            var bounds = ambulance.getBounds();
-            createjs.Tween.get(ambulance).to({
-                x: ((ambulance.x + bounds.width) < w) ? ambulance.x + 10 : ambulance.x
-            }, 30);
-        }
-
-
-        if (e.keyCode === 37) {
-            var bounds = ambulance.getBounds();
-            createjs.Tween.get(ambulance).to({
-                x: (ambulance.x > 0) ? ambulance.x - 10 : ambulance.x
-            }, 30);
-        }
-
-    }
-}
-
-$(window).on('keydown', moveSprite);
